# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Code Editor & Compiler API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** Primary language (html, css, javascript, react, python, typescript, etc.) */
  language: string;
  fileCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface CodeFile {
  id: number;
  projectId: number;
  name: string;
  language: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectWithFiles {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  language: string;
  files: CodeFile[];
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  language: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  language?: string;
}

export type ProjectStatsByLanguageItem = {
  language: string;
  count: number;
};

export interface ProjectStats {
  totalProjects: number;
  totalFiles: number;
  /** Projects updated in last 7 days */
  recentActivity: number;
  byLanguage: ProjectStatsByLanguageItem[];
}

export interface CodeFileInput {
  /** @minLength 1 */
  name: string;
  language: string;
  content: string;
}

export interface CodeFileUpdate {
  /** @minLength 1 */
  name?: string;
  language?: string;
  content?: string;
}

export interface RunCodeInput {
  language: string;
  code: string;
  stdin?: string;
}

export interface RunCodeResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  /** @nullable */
  executionTime?: number | null;
}