# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Code Editor & Compiler API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Code project operations
  - name: files
    description: File operations within projects
  - name: execution
    description: Code execution operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/recent:
    get:
      operationId: listRecentProjects
      tags: [projects]
      summary: List recently modified projects
      responses:
        "200":
          description: Recently modified projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /projects/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get project statistics
      responses:
        "200":
          description: Aggregated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with its files
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project with files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWithFiles"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{id}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Files list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodeFile"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeFileInput"
      responses:
        "201":
          description: Created file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeFile"

  /projects/{id}/files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get a specific file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeFile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFile
      tags: [files]
      summary: Update a file's content or name
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeFileUpdate"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeFile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /run:
    post:
      operationId: runCode
      tags: [execution]
      summary: Execute code server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCodeInput"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunCodeResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Project:
      type: object
      required: [id, name, language, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
          description: Primary language (html, css, javascript, react, python, typescript, etc.)
        fileCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    ProjectWithFiles:
      type: object
      required: [id, name, language, createdAt, updatedAt, files]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/CodeFile"
        createdAt:
          type: string
        updatedAt:
          type: string

    ProjectInput:
      type: object
      required: [name, language]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        language:
          type: string

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        language:
          type: string

    ProjectStats:
      type: object
      required: [totalProjects, byLanguage, totalFiles, recentActivity]
      properties:
        totalProjects:
          type: integer
        totalFiles:
          type: integer
        recentActivity:
          type: integer
          description: Projects updated in last 7 days
        byLanguage:
          type: array
          items:
            type: object
            required: [language, count]
            properties:
              language:
                type: string
              count:
                type: integer

    CodeFile:
      type: object
      required: [id, projectId, name, language, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        language:
          type: string
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    CodeFileInput:
      type: object
      required: [name, language, content]
      properties:
        name:
          type: string
          minLength: 1
        language:
          type: string
        content:
          type: string

    CodeFileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        language:
          type: string
        content:
          type: string

    RunCodeInput:
      type: object
      required: [language, code]
      properties:
        language:
          type: string
        code:
          type: string
        stdin:
          type: string

    RunCodeResult:
      type: object
      required: [stdout, stderr, exitCode]
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        executionTime:
          type: ["number", "null"]