# lib/api-zod/src/generated/types/codeFile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Code Editor & Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface CodeFile {
  id: number;
  projectId: number;
  name: string;
  language: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}