# lib/api-zod/src/generated/types/codeFileUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Code Editor & Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface CodeFileUpdate {
  /** @minLength 1 */
  name?: string;
  language?: string;
  content?: string;
}