# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Code Editor & Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** Primary language (html, css, javascript, react, python, typescript, etc.) */
  language: string;
  fileCount?: number;
  createdAt: string;
  updatedAt: string;
}