# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Code Editor & Compiler API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatsByLanguageItem } from './projectStatsByLanguageItem';

export interface ProjectStats {
  totalProjects: number;
  totalFiles: number;
  /** Projects updated in last 7 days */
  recentActivity: number;
  byLanguage: ProjectStatsByLanguageItem[];
}