# lib/api-zod/src/generated/types/projectWithFiles.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Code Editor & Compiler API
 * OpenAPI spec version: 0.1.0
 */
import type { CodeFile } from './codeFile';

export interface ProjectWithFiles {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  language: string;
  files: CodeFile[];
  createdAt: string;
  updatedAt: string;
}