# lib/api-zod/src/generated/types/runCodeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Code Editor & Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface RunCodeInput {
  language: string;
  code: string;
  stdin?: string;
}