# lib/api-zod/src/generated/types/runCodeResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Code Editor & Compiler API
 * OpenAPI spec version: 0.1.0
 */

export interface RunCodeResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  /** @nullable */
  executionTime?: number | null;
}